#507 add new components wrapper
